move lazyload subscriber tests
